DEV-5555
Title change to "Payment Gateway – nexi Alpha Bank for WooCommerce"
Major code cleanup
Removed backward compatibilities to make future development easier
