--- woo-alpha-bank-payment-gateway/encryption.php
@@ -1,41 +1,0 @@
-<?php
-class WC_Payment_Gateway_KeyEncryption_Alpha
-{
-    const METHOD = 'aes-128-cbc';    
-    public static function encrypt($message, $key)
-    {
-        if (mb_strlen($key, '8bit') !== 32) {
-            throw new Exception("Needs a 256-bit key! ".mb_strlen($key, '8bit'));
-        }
-        $ivsize = openssl_cipher_iv_length(self::METHOD);
-        $iv = openssl_random_pseudo_bytes($ivsize);
-        
-        $ciphertext = openssl_encrypt(
-            $message,
-            self::METHOD,
-            $key,
-            0,
-            $iv
-        );
-        
-        return $iv . $ciphertext;
-    }
-
-    public static function decrypt($message, $key)
-    {
-        if (mb_strlen($key, '8bit') !== 32) {
-            throw new Exception("Needs a 256-bit key! ".mb_strlen($key, '8bit'));
-        }
-        $ivsize = openssl_cipher_iv_length(self::METHOD);
-        $iv = mb_substr($message, 0, $ivsize, '8bit');
-        $ciphertext = mb_substr($message, $ivsize, null, '8bit');
-        
-        return openssl_decrypt(
-            $ciphertext,
-            self::METHOD,
-            $key,
-            0,
-            $iv
-        );
-    }
-}
